// src/redux/enums/UsuarioEnum.ts
export enum UsuarioEnum {
   POPULATE_USUARIO_START = 'POPULATE_USUARIO_START',
   POPULATE_USUARIO_SUCCESS = 'POPULATE_USUARIO_SUCCESS',
   POPULATE_USUARIO_ERROR = 'POPULATE_USUARIO_ERROR',

   ADD_USUARIO_START = 'ADD_USUARIO_START',
   ADD_USUARIO_SUCCESS = 'ADD_USUARIO_SUCCESS',
   ADD_USUARIO_ERROR = 'ADD_USUARIO_ERROR',

   DELETE_USUARIO_START = 'DELETE_USUARIO_START',
   DELETE_USUARIO_SUCCESS = 'DELETE_USUARIO_SUCCESS',
   DELETE_USUARIO_ERROR = 'DELETE_USUARIO_ERROR'
}